requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
mlflow>=2.7.0
dvc>=3.0.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
crate>=0.3.0
pymongo>=4.5.0
pyyaml>=6.0
boto3>=1.28.0
pytest>=7.4.0
schedule>=1.2.0
psutil>=5.9.0
uvicorn>=0.20.0
fastapi>=0.95.0
celery>=5.3.0
redis>=4.5.0
streamlit>=1.25.0
